### 2016-06-25 #56 @daattali
ezknitr package
* **1. What does this package do? (explain in 50 words or less)**

Allows you to use 'knitr' with complex directory structures and gives you more control over where inputs/outputs/figures/scripts live when knitting documents. Real research projects often don't have a flat directory structure, which can make using 'knitr' on such projects very cumbersome.

* **2. Paste the full DESCRIPTION file inside a code block below.**

```
Package: ezknitr
Title: Avoid the Typical Working Directory Pain When Using 'knitr'
Version: 0.4
Authors@R: person("Dean", "Attali", email = "daattali@gmail.com",
    role = c("aut", "cre"))
Description: An extension of 'knitr' that adds flexibility in several
    ways. One common source of frustration with 'knitr' is that it assumes
    the directory where the source file lives should be the working directory,
    which is often not true. 'ezknitr' addresses this problem by giving you
    complete control over where all the inputs and outputs are, and adds several
    other convenient features to make rendering markdown/HTML documents easier.
    knitr (>= 1.7),
    markdown (>= 0.7),
    R.utils (>= 1.34.0)
SystemRequirements: pandoc with https support
```

* **3. URL for the package (the development repository, not a stylized html page)**

https://github.com/daattali/ezknitr

* **4. What data source(s) does it work with (if applicable)?**

N/A

* **5. Who is the target audience?**

Any researcher or grad student who wants to better automate and organize their research using `knitr` but find it annoying/impossible to deal with directories in a sane way

* **6. Are there other R packages that accomplish the same thing? If so, what is different about yours?**

    - **rmarkdown** has support for parameters in Rmd documents.  One of the features that `ezknitr` offers is the ability to pass parameters to Rmd files. There is a difference between the two: rmarkdown assumes that the parameters are defined in the Rmd YAML, while ezknitr lets you change any arbitrary variable. This difference is not too significant, but the main reason ezknitr has this feature is to support older Rmd files that don't have YAML because the introduction of parameterized Rmd files is fairly new

    - **rprojroot** provides a very different solution to a related problem. `rprojroot` helps you find a specific file in relation to a project's root directory, while ezknitr is specifically for dealing with knitting documents in complex directory structures

* **7. Check the box next to each policy below, confirming that you agree. These are mandatory.**
    * [x] This package does not violate the Terms of Service of any service it interacts with.
    * [x] The repository has continuous integration with Travis CI and/or another service
    * [x] The package contains a vignette
    * [x] The package contains a reasonably complete README with instructions for installing the development version.
    * [x] The package contains unit tests
    * [x] The package only exports functions to the NAMESPACE that are intended for end users

* **8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.**
    * [ ] Are there any package dependencies not on CRAN?
    * [x] Does the package have a CRAN accepted license?
    * [ ] Did `R CMD check` (or `devtools::check()`) produce any errors or warnings? If so paste them below.

Yes, I will follow those guidelines.  The package is already on CRAN.

* **9. Please add explanations below for any exceptions to the above:**

N/A

* **10. If this is a resubmission following rejection, please explain the change in circumstances.**

N/A

* **UPDATE 11. JOSS publication**

    - [x] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description.

        > yes, I included `paper.md` under `/inst/joss/`

        > figshare [10.6084/m9.figshare.3545363](https://dx.doi.org/10.6084/m9.figshare.3545363)
