### 2016-06-27 #56 @noamross
Thanks for the submission, @daattali. I'll be your handling editor.  I'll do editor's checks this week before assigning reviewers.
