### 2016-06-27 #56 @noamross
Hi @daattali, could you comment on Question 6 above re: overlapping packages?  As you note in the README, there's some feature overlap with **rmarkdown**. Also **rprojroot** provides an alternative solution for the working directory issue.  Neither of these preclude acceptance of **ezknitr**, but we'd like your take on the advantages/disadvantages of your approach and which use-cases it is most appropriate for, which will help reviewers give feedback.
