### 2016-06-28 #56 @daattali
@noamross you're absolutely right, I edited Q6
