### 2016-07-05 #56 @noamross
Reviewers: @karthik @ttimbers
Due date: 2016-07-21
