### 2016-07-20 #56 @ttimbers
## Package Review and Comments

Overall this package looks really good. It solves an important limitation of the `knitr` package; managing knitting and spinning files located in different directories, as well as providing the flexibility and control over where the output files are saved to. Everything worked well on my machine, and my review and comments are mainly in respect to improving documentation and style to match what is expected of rOpenSci packages.

#### Documentation comments

- To make the README.md even more useful for all audiences (especially for those who are newer to R) it would be useful to have a 1-2 sentences in the first paragraph to describe the purpose/role of `knitr` in the README.md before `ezknitr` is defined as an extension of `knitr`. It would also be good to link to the `knitr` package documentation from your README.md file when you reference it.

- To access the `setup_ezknit_test` or `setup_ezspin_test` example or documentation within R you have to load the library via `library(ezknitr)` after installing to get `setup_ezknit_test()`, `setup_ezspin_test()`, or `?setup_ezknit_test` to work. This should be clarified in the documentation.

- In general, the examples are really good, and clearly illustrate how you can use the package. They all ran locally, without error, on my machine.

- It would be nice to have a function or provide a command to the user to clean up after each of the examples in `setup_ezknit_test()` and `setup_ezspin_test()`. This would make it easier for the user to understand what the different commands do.

- There is a vignette associated with the package, located in the vignettes directory, but I could not access it locally within R using `browseVignettes("ezknitr")` or by using `vignette(package = "ezknitr")` within R. I can also access the vignette through this link on CRAN: https://cran.r-project.org/web/packages/ezknitr/vignettes/overview.html. If not too much work, it would be nice to have the vignette appear within R using both `browseVignettes("ezknitr")` and `vignette(package = "ezknitr")`.

- rOpenSci footer image is not yet included (this is my first time reviewing a package, so maybe this need not be included until after acceptance?)

- There is no `README.Rmd` file, only a `README.md` file in the repository on Github, which leads me to believe the README.md was created directly, and not via rendering of an `README.Rmd` file. rOpenSci recommends creating a README.md from a README.Rmd file where possible because.

- Currently there is no documentation on contribution guidelines. These should be included in either the `README` or in a file called `CONTRIBUTING`. Additionally, although the `URL` and `BugReports` fields are present in `DESCRIPTION`, the `Maintainer` field is missing. There is also no Code of Conduct. rOpenSci recommends that you use a code of conduct such as the [Contributor Covenant](http://contributor-covenant.org/) in developing your package.

- Although the package uses semantic versioning on CRAN, there are no release tags for these versions on Github. Please add these.

#### Functionality comments

- The code appears to be efficient, and the usage of each function is clear. There is no code duplication in the package that should be reduced.

- I might be tempted to split the `ezknitr_helper()` function into at least two functions, one that figures out all the paths, and one that uses that information to call the `knitr::spin()` or `knitr::knit()` function. This is not essential, but it might make future development easier (*e.g.,* if you decide in the future to set the paths some other way).

- In general, edge cases have been considered for most functions. In the function `ezknitr_helper()`, the edge case of someone passing `out_suffix` a value indexed outside of a list is not considered. For example, if you had a list called `mylist` of length 3, and you set `out_suffix = mylist[4]` then suffix would become `NA`. I wonder if this is a desired behaviour?

- There is a good coverage of tests (a reasonable range of inputs and conditions) using `testthat` for the two core functions (`ezspin()` and `ezknit()`) and the `set_default_params()` function.

- There are no tests for the `open_output_dir()` or `set_default_params()` functions. `open_output_dir()` is not an essential function, and so maybe it is OK that there are no tests for it?

- Although much of the code implements `snake_case`, there are some occurrences of variable names using `lowerCamelCase`. Examples from `ezknitr_helper()` include: `fileName`, `fileRmdOrig`, `fileRmd`, `fileMd`, `fileHtml`, `oldOptsKnit`, `oldOptsChunk`, and `fullFigPath`.

#### Future functionality ideas:

- It would be great if reports were also generated in PDF. One reason that this would be is useful is that although html renders LaTex, html does not render on Github. Markdown, renders nicely on Github, but it does not render the embedded LaTeX. PDFs are great because they do both. Additionally, PDFs are great documents to share with non-coding collaborators.

- How might this interface with the new R Notebooks? Apologies for the open-ended question in a review, but I am curious and can't help asking.

---

Estimated hours spent reviewing: 4 hrs

---
