### 2016-07-20 #56 @daattali
Thanks @ttimbers , great and thorough review!  I'll wait until @karthik submits his review to respond and take action
