### 2016-08-01 #56 @sckott
@karthik @ttimbers
Due date: 2016-07-21 - hey there, it's been 27 days, please get your review in soon, thanks :smiley_cat: (ropensci-bot)
