### 2016-08-06 #56 @ropenscibot
@karthik @ttimbers
Due date: 2016-07-21 - hey there, it's been 32 days, please get your review in soon, thanks :smiley_cat: (ropensci-bot)
