### 2016-08-06 #56 @daattali
Can I retroactively add the "automatic publishing to JOSS" info?
