### 2016-08-07 dab0fcc
add copyright and license text to readme
