+++ vignettes/overview.Rmd
@@ -18,11 +18,11 @@ knitr::opts_chunk$set(tidy = FALSE, comment = "#>")
 [![Build Status](https://travis-ci.org/daattali/ezknitr.svg?branch=master)](https://travis-ci.org/daattali/ezknitr)
 [![CRAN version](http://www.r-pkg.org/badges/version/ezknitr)](https://cran.r-project.org/package=ezknitr)
 
+> *Copyright 2016 [Dean Attali](http://deanattali.com). Licensed under the MIT license.*
+
 `ezknitr` is an extension of `knitr` that adds flexibility in several ways.
 
 One common source of frustration with `knitr` is that it assumes the directory where the source file lives should be the working directory, which is often not true. `ezknitr` addresses this problem by giving you complete control over where all the inputs and outputs are, and adds several other convenient features. The two main functions are `ezknit()` and `ezspin()`, which are wrappers around `knitr`'s `knit()` and `spin()`, used to make rendering markdown/HTML documents easier. 
-
-> You can see Jenny Bryan's way of dealing with this problem [in this gist](https://gist.github.com/jennybc/362f52446fe1ebc4c49f) or simply browse the [knitr GitHub issues](https://github.com/yihui/knitr/issues) to see people discussing the issue surrounding directories.
 
 ## Table of contents
 
