### 2016-08-07 41b9581
add JOSS paper
