+++ inst/joss/paper.bib
@@ -1,0 +1,36 @@
+@misc{ezknitr-archive,
+  author       = {Dean Attali,
+  title        = {ezknitr: Avoid the Typical Working Directory Pain When Using 'knitr'},
+  month        = jul,
+  year         = 2016,
+  doi          = {10.6084/m9.figshare.3545363},
+  url          = {https://dx.doi.org/10.6084/m9.figshare.3545363.v1}
+}
+
+@Manual{R-base,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2016},
+  url = {https://www.R-project.org/},
+}
+
+@Manual{R-knitr,
+  title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
+  author = {Yihui Xie},
+  year = {2014},
+  note = {R package version 1.13.7},
+  url = {https://CRAN.R-project.org/package=knitr},
+}
+
+@article{knuth1984literate,
+  title={Literate programming},
+  author={Knuth, Donald Ervin},
+  journal={The Computer Journal},
+  volume={27},
+  number={2},
+  pages={97--111},
+  year={1984},
+  publisher={Br Computer Soc}
+}
