### 2016-08-07 bd92bcb
add paper.md to JOSS paper
