### 2016-08-07 #56 @daattali
I have updated the [initial post](https://github.com/ropensci/onboarding/issues/56#issue-162274746) in this thread and added an 11th bullet point mentioning that I want to submit to JOSS.  Perhaps @ttimbers and @karthik can take a quick look at that and give their opinion on that as well?  I believe it should take no longer than a few minutes. Thanks!
