### 2016-08-12 #56 @karthik
# Review of ezknitr

This is an extremely useful addition to the `knitr` reproducible research workflow and addresses many frustrations that people have when working with projects where all assets are not self-contained.
The package installs with no issues and all tests pass correctly when I reviewed at `bd92bcb`. Also not duplicating any comments from @ttimbers above.

**Comments**

- `setup_ezspin_test` , `setup_test_helper`

I was going to suggest having a few example use cases, such as a simple report with one level nesting, a complex report (or multiple reports) with data in various places. But the `setup_ezknit_test()` is a nice addition. Perhaps the function could take two options (one for simple and one for complex) and spit out two workflows? Take or leave this suggestion (i.e. I would not impose it as a requirement to acceptance) but including this as a suggestion.

- `set_default_params`

> ' Sometimes it may be useful to define a variable only it hasn't been defined yet.
> ' One example where this can be useful is when you have an Rmd script that
> ' uses some variables and you want to be able to use custom values for these
> ' variables, but also give them a default value in the script in case they are
> ' not set beforehand.

The function itself does not check to see if these variables have been defined in a document. If I understand correctly, you are asking the user to use `exists` to first check, then set these as defaults (which could be overwritten by passing new args, right?). If so, couldn't the function itself check to see if those variable s exist and throw a warning/error if not (which could be overridden)?

```r
set_default_params <- function(params) {
  params <- as.list(params)
  env <- parent.frame(1)

  invisible(
    lapply(names(params), function(key) {
      if (!exists(key, envir = env, inherits = FALSE)) {
      # ^ Is this not supposed to work outside of ezknit?
        assign(key, params[[key]], envir = env, inherits = FALSE)
      }
    })
  )
}
```

Also, rather than point out that `rmarkdown` has solved this issue with yml further down, perhaps you could move that up here?

All other functionality looks as expected.

**Other suggestions for improvement**

> Now if you want knitr to work, you'd have to ensure the path to input.csv is relative to the analysis/ directory. This is counter-intuitive because most people expect to create paths relative to the working directory (project/ in this case), but knitr will use the analysis/ folder as the working directory. Any code reading the input file needs to use ../data/input.csv instead of data/input.csv.

This came off as a bit confusing to me at first. Perhaps you could say _project root_ to keep it more general?

> We didn't explicitly have to set the working direcory, but you can use the wd argument if you do require a different directory.

Perhaps you can say why this might be useful to do. E.g. if someone is doing this from a makefile or another build script, this could be the way to not worry about where that file resides (and `wd` will be set explicitly.)

> Let's assume that each analysis script expects there to be a variable named DATASET_NAME that tells the script what data to operate on.

When reading the README, it is not clear if the variable DATASET_NAME would be set in the yml or the knitr document itself. It is the latter, but it would be good to spell out and perhaps even include a little code snippet.

> knit() is the most popular and well-known function from knitr. It lets you create a markdown document from an Rmarkdown file. I assume if you are on this page, you are familiar with knit() and Rmarkdown, so I won't explain it any further.

Suggest linking out to Yihui's page in case such a person is reading.

# Overall comments

I think this is a fantastic addition to the rOpenSci suite and recommend acceptance. 💯🚀

![](https://dl.dropboxusercontent.com/u/2223411/gifs/howidointernet.gif)
