### 2016-08-12 #56 @noamross
Thanks @karthik! Could you prepend the [reviewer checklist](
https://github.com/ropensci/onboarding/blob/master/reviewer_template.md) to
your review, and also comment on `paper.md`? This will likely be our first
test of the JOSS pass-through.
