### 2016-08-21 13b9ac4
revisions based on ropensci feedback
