+++ inst/joss/paper.md
@@ -13,6 +13,6 @@ bibliography: paper.bib
 
 # Summary
 
-The 'ezknitr' package [@ezknitr-archive] is an R package [@R-base] that enhances the functionality and flexibility of the 'knitr' package [@R-knitr]. 'knitr' is a popular package for generating dynamic reports in R using the concept of literate programming [@knuth1984literate]. One common source of frustration with 'knitr' is that it makes important assumptions about the directory structure of the input files and does not offer much flexibility in deciding where the different output files are generated. The 'ezknitr' package adresses these problems by providing the user with complete control over where all the inputs and outputs are, and adds several other convenient features to make the process of generating dynamic documents more streamlined and efficient.
+'knitr' [@R-knitr] is a popular R package [@R-base] that implements the concept of literate programming [@knuth1984literate] in R. It has been widely adopted in the R community as a means of generating dynamic reports. Despite its popularity, one common source of frustration with 'knitr' is that it makes important assumptions about the directory structure of the input files and does not offer much flexibility in deciding where the different output files are generated. The 'ezknitr' package [@ezknitr-archive] is an R package that addresses these problems and enhances the functionality of 'knitr'. 'ezknitr' provides the user with complete control over where all the inputs and outputs are, and adds several other convenient features to make the process of generating dynamic documents more streamlined and efficient.
 
 # References
