+++ vignettes/overview.R
@@ -1,0 +1,3 @@
+## ----setup, echo = FALSE, message = FALSE--------------------------------
+knitr::opts_chunk$set(tidy = FALSE, comment = "#>")
+
