### 2016-08-21 #56 @daattali
Thanks for taking your time to review this  @karthik and @ttimbers

I've updated ezknitr based on all your comments, let me know if you have any further comments
