### 2016-08-25 #56 @daattali
Sorry, I just noticed that I forgot to add all my comments (I had them sitting in a text file -- it 'd be awesome if GitHub supported "saving drafts" of comments).  I added/improved everything that was suggested, but here are some comments about things that I didn't change:

### @ttimbers
I thought about adding a cleanup function but decided against it because depending on what function the user runs, files and folders will get generated in different places and they can run any arbitrary command so I can't know what exactly to clean without making a lot of assumptions

browseVignette: I think if you install the package from CRAN or GitHub then you will see the vignettes, it's only when you install using devtools that the vignettes don't get built/linked https://github.com/hadley/devtools/issues/931

README.Rmd - I chose to stick with just a direct README.md because I don't have code chunks. If in the future I'll have R code chunks that will benefit from Rmd -> md then I'll change it

out_suffix: if you try to assign it an NA value , then the function will fail with an informative error message ("Error: `out_suffix` is not a valid string"). I believe this is the correct behaviour. Maybe you meant something else?

open_output_dir has no test because all it does is a side effect that is (impossible? super extremely hard?) to test. I don't know of any way to test that a function caused the OS file explorer to open. set_default_params() does have tests https://github.com/daattali/ezknitr/blob/master/tests/testthat/test-set_default_params.R

compiling to PDF: as far as I know, there isn't any special function to do that, right? You just need to use pandoc? I'm open to a pull request or an issue being opened but I don't have any experience making PDF so I prefer to leave it as a future improvement.

I haven't actually played around with notebooks too much, but I don't think it will be able to interact with notebooks because notebooks seem like they use knitr under the hood (I think - I don't actually know its internals) which means they'll still have the usual work directory problems

### @karthik

set_default_params: the user does not need to make any checks themselves. All the user does is pass in a named list, and every key in the list will be treated as a variable name. Every variable name that currently doesn't exist will get created and assigned the value from the list. The function doesn't need to explicitly check for existence using "exists()" because I do use exists() inside the loop structure to see if an incoming variable exists or not.

set_default_params WILL work in any context, not just in ezknitr. you can try it yourself: assign 'x <- 5', then call 'set_default_params(list(x=10, y=20))'. Hopefully x will remain untouched but y will be 20

PS. SO COOL that you can have a little Details bar that collapses/expands on click!
