### 2016-09-12 #56 @ttimbers
@daattali, my comments to your comments:

1. Fair enough justification for not adding a clean-up function.

2. You are correct (and thanks for informing me!) that `vignette` and `browseVignette` work only when packages are installed from CRAN. When I install `ezknitr` from CRAN, and then run `browseVignettes("ezknitr")` I get directed to a useful webpage where I can access a link ("HTML") to the vignetter (although it might be better to just be taken to the HTML version of the vignette as the other links are not very useful - if this is possible). On the other hand, `vignette(package = "ezknitr")` doesn't seem very helpful in its current implementation. When run is outputs what is pasted below to the editor.

I know that these calls are using an older version of the `ezknitr` package (the one on CRAN), but presumably you will put this reviewed version of the package on CRAN and it would be great to have `vignette(package = "ezknitr")`  functional.

```
Vignettes in package ‘ezknitr’:

overview                     Package ezknitr (source, html)
```

3. As long as ROpenSci is OK with you using a direct README.md, that is fine with me. I commented on it, as they recommend the use of README.Rmd knit to README.md over a direct README.md. @noamross maybe you can comment on how strict ROpenSci is with this?

4. OK that `out_suffix` is giving you the desired output. I just wanted to check.

5. Fair enough that `open_output_dir` has no test because it is too difficult to do. Also, apologies that I missed the tests for `set_default_params`, I now see that they exist.

6. I do still think that because users are used to being able to `knit` to PDF with `knitr`, that users might want and expect this functionality from `ezknitr`. Especially users who are used to only working in R, and have never run command line programs, such as pandoc. I also appreciate that this may be beyond the scope of the current development stage of the `ezknitr` package and can be left for future development.

7. Again, I appreciate integrating `ezknitr` with RStudio notebooks might be beyond the current scope of the `ezknitr` package, but thought I would mention it, as I could envision how they might function well together.
