### 2016-09-12 #56 @ttimbers
### Review of the JOSS paper

Looks good to me! I recommend ROpenSci accept the paper and the package. 😃
