### 2016-09-12 #56 @noamross
Using a plain README.md is fine if there are no code chunks.  I've just clarified the language in the packaging guide
