### 2016-09-12 #56 @noamross
Approved!  This will be our first JOSS pass-through package to reach this stage of the process.  Allow me to pre-apologize for any snafus along the way.  Here are the remaining steps, @daattali:

-   Close any other outstanding issues if you'd like. (I'll do a final editor's check if you choose to include markdown or PDF output in this release.)
-   Add the rOpenSci footer to the README
-   Transfer to the repo to rOpenSci
-   Update the figshare repo to match the ropensci-accepted version.
-   Fill out the form at http://joss.theoj.org/papers/new to submit to JOSS.  We will flag the package to let the editor know they may accept based on this review history.
