### 2016-09-13 #56 @daattali
@ttimbers

I think the behaviour you're seeing from both `browseVignettes("ezknitr")` and `vignette(package = "ezknitr")` is unfortunately the correct behaviour of those functions, even if it doesn't seem super useful.  I tried replacing "ezknitr" with "tidyr" to verify, and I get the same results

Re: knitting to PDF: I would happily add support for ezknit to pdf, but does knitr::knit have support for that in the first place? I never do PDF myself so I don't know how people do it. I tried googling it a bit and as far as I can tell, knitr does not have a way to convert to PDF, only `rmarkdown::render` does, which is independent from knitr.

Can you let me know if you do know of a way to convert to PDF using knitr?  If not, then I'll treat this package as done for now and will go through with the next steps
