### 2016-09-14 #56 @ttimbers
Re: `browseVignettes` and `Vignettes`

`tidyr` is not a great example, as it does not work with `vignette`, I am guessing because the developers didn't implement it. Both `stringr` and `broom` work well with `vignette`.

===

Re: knitting to PDF

I guess RStudio had fooled me into thinking the "Knit PDF" button uses a function from the `knitr` package as opposed to the `rmarkdown` package...
