### 2016-09-14 #56 @ttimbers
Re: Re: knitting to PDF

It looks like in theory, you could use `knitr` to do this via `knitr::knit` to convert `.Rmd` to `.Rrst` or `.Rnw`, and then `knitr::knit2pdf` to convert the `.Rrst` or `.Rnw` to `.pdf`.

Although I must say, I tried this strategy and it didn't work as I ran into an error with `rst2pdf` (a required dependency for `knitr::knit2pdf`?).

Anyways, its a bit frustrating that it is not obvious to if this is what the "Knit PDF" button is doing, and if not why its says "Knit" if it is not using `knitr`...
