+++ Makefile
@@ -4,19 +4,19 @@
 all: README.md
 
 clean:
-	Rscript -e 'suppressWarnings(file.remove("README.md", "vignettes/overview.md"))'
+	Rscript -e 'suppressWarnings(file.remove("README.md", "vignettes/ezknitr.md"))'
 
 .PHONY: all clean
 .DELETE_ON_ERROR:
 .SECONDARY:
 
-README.md : vignettes/overview.Rmd
-#	echo "Rendering the overview vignette"
-	Rscript -e 'rmarkdown::render("vignettes/overview.Rmd", output_format = "md_document")'
+README.md : vignettes/ezknitr.Rmd
+#	echo "Rendering the ezknitr vignette"
+	Rscript -e 'rmarkdown::render("vignettes/ezknitr.Rmd", output_format = "md_document")'
 #	echo "Correcting image paths"
-#	sed -i -- 's,../inst,inst,g' vignettes/overview.md
-	Rscript -e 'file <- gsub("../inst", "inst", readLines("vignettes/overview.md")); writeLines(file, "vignettes/overview.md")'
+#	sed -i -- 's,../inst,inst,g' vignettes/ezknitr.md
+	Rscript -e 'file <- gsub("../inst", "inst", readLines("vignettes/ezknitr.md")); writeLines(file, "vignettes/ezknitr.md")'
 #	echo "Copying output to README.md"
-#	cp vignettes/overview.md README.md
-	Rscript -e 'file.copy("vignettes/overview.md", "README.md", overwrite = TRUE)'
-	Rscript -e 'suppressWarnings(file.remove("vignettes/overview.md"))'
+#	cp vignettes/ezknitr.md README.md
+	Rscript -e 'file.copy("vignettes/ezknitr.md", "README.md", overwrite = TRUE)'
+	Rscript -e 'suppressWarnings(file.remove("vignettes/ezknitr.md"))'
