### 2016-09-14 #56 @daattali
Ohh I see what you mean about vignettes! I was under the impression you were trying to call `vignette(package = "tidyr")` , which *lists* its vignettes. You mean just calling `vignette("stringr")`, which shows the actual vignette. That just means that stringr has a vignette named `stringr`. To see the ezknitr vignette like that you would use `vignette("overview", package = "ezknitr")`, is that what you were trying to get at? Anyway, you're right, I never thought much about the name I was giving vignettes. I will change the vignette name to `ezknitr` so that it will work as you expected, and I'll also start doing it for my other packages :)

Yes the button is a bit misleading, I'm pretty sure it uses rmarkdown::render...
