### 2016-09-14 #56 @daattali
@noamross I can't transfer the repo to rOpenSci, it tells me I need admin right
