### 2016-09-14 #56 @daattali
I was just invited to ropenscilabs. Am I supposed to transfer ownership to ropenscilabs or to ropensci?

If it's the former, then will it eventually move to ropensci? Does that mean I should update all the links now to point to github.com/ropenscilabs/ezknitr and later on I'll have to update again to point to ropensci?
