### 2016-09-14 #56 @daattali
Sorry if this info is available elsewhere, I couldn't find it anywhere.  Do packages move from ropenscilabs to ropensci in a matter of hours/days, or do they remain in labs for a long time?  (So that I can maybe just skip the ropensci labs URL if it's only temporary)
