### 2016-09-14 #56 @noamross
They remain for a long time. Honestly there's little difference between the
two orgs. We anticipated at one point there being a distinction based on
age/stability of packages but that process hasn't really evolved.
Thankfully all the URLs do forward so when it's transfered (or we
re-consolidate) the links will work.
