### 2016-09-14 #56 @daattali
I'll do a CRAN submission and when the new version gets on CRAN I'll continue with the transfer process
