### 2016-09-14 8fbee2c
Update cran-comments.md
