### 2016-09-14 #56 @karthik
>  Do packages move from ropenscilabs to ropensci in a matter of hours/days

Not a matter of hours. After the project grew and the main account was getting cluttered with mature projects and various new/half-finished/non-software repos, we started ropenscilabs. All new projects started by the rOpenSci team begin there, and all accepted projects from the community (like yours) get transferred from a personal account to there.

Like @noamross said, GitHub does a great job forwarding repos as they move around. We'll update the `BugReports` link when it moves later but for now it will reside in labs. We will clarify the policy.
