### 2016-09-14 915f8ec
Update cran-comments.md
