+++ cran-comments.md
@@ -60,3 +60,7 @@ Thanks, on CRAN now.
 No errors, warnings, or notes
 
 ### Reviewer comments
+
+2016-09-14 Kurt Hornik
+
+Thanks, on CRAN now
