### 2016-09-14 17b0a5b
update daattali to ropenscilabs in vignette
