+++ vignettes/ezknitr.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(tidy = FALSE, comment = "#>")
 
 # ezknitr - Avoid the typical working directory pain when using 'knitr'
 
-[![Build Status](https://travis-ci.org/daattali/ezknitr.svg?branch=master)](https://travis-ci.org/daattali/ezknitr)
+[![Build Status](https://travis-ci.org/ropenscilabs/ezknitr.svg?branch=master)](https://travis-ci.org/ropenscilabs/ezknitr)
 [![CRAN version](http://www.r-pkg.org/badges/version/ezknitr)](https://cran.r-project.org/package=ezknitr)
 
 > *Copyright 2016 [Dean Attali](http://deanattali.com). Licensed under the MIT license.*
@@ -48,7 +48,7 @@ To install the latest developmental version from GitHub:
 
 ```
 install.packages("devtools")
-devtools::install_github("daattali/ezknitr")
+devtools::install_github("ropenscilabs/ezknitr")
 ```
 
 <h2 id="overview">Overview</h2>
