### 2016-09-14 #56 @daattali
I just transferred the repo and have a few comments and a question:

- It wasn't clear to me that when transferring I should see all the teams that exist in the org and choose the `ezknitr` team only (I assume that was the correct thing to do?)
- I'm not sure why nobody else ran into this problem before, but my travis build failed once I transferred ownership because the travis badge URL and the travis build URL had to be updated to replace `daattali` with `ropenscilabs`. So even though github does an awesome job of redirecting, it seems like other services do not
- in the JOSS submission, in the field that asks for "Repository address", should I use the GitHub repo URL or the figshare URL?
