### 2016-09-15 #56 @daattali
@noamross in the JOSS submission, for the field of "Repository address", do I enter the github repo or the figshare url?
