### 2016-09-15 #56 @noamross
1) This was the correct thing to do, thanks!
2) I'm not sure how often this happens.  We need to turn on @sckott?
3) It should be the github URL.

Thanks!  We're collecting the notes from these little confusions for our first few JOSS submissions.
