### 2016-09-15 #56 @noamross
Nope! All done. I'll close this issue and keep you appraised. Expect some notifications from the JOSS side.
