### 2016-09-15 #56 @sckott
> I'm not sure how often this happens. We need to turn on?

yeah, I think we need to manually turn on builds
